taken out create and reset buttons into its own component
