+++ client/src/components/Visualiser/Controller/VisualiserControls.tsx
@@ -21,6 +21,7 @@ import React, { EventHandler, FC, useCallback, useContext, useEffect, useState }
 import { defaultSpeed } from 'visualiser-src/common/constants';
 import VisualiserContext from '../VisualiserContext';
 import styles from './Control.module.scss';
+import VisualiserCreateNewReset from './VisualiserCreateNewReset';
 
 /**
  * Contains all the visualiser controller UI, ie. the play/pause buttons, the
@@ -74,14 +75,6 @@ const VisualiserControls = () => {
     },
     [controller]
   );
-
-  const handleReset = useCallback(() => {
-    controller.resetDataStructure();
-  }, [controller]);
-
-  const handleGenerate = useCallback(() => {
-    controller.generateDataStructure();
-  }, [controller]);
 
   const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
     setAnchorEl(event.currentTarget);
@@ -215,16 +208,7 @@ const VisualiserControls = () => {
           }}
         />
       </Box>
-      <Button className={styles.resetButton} onClick={handleGenerate}>
-        <Typography color="textPrimary" sx={{ whiteSpace: 'nowrap' }}>
-          Create New
-        </Typography>
-      </Button>
-      <Button className={styles.resetButton} onClick={handleReset}>
-        <Typography color="textPrimary" whiteSpace="nowrap">
-          Reset All
-        </Typography>
-      </Button>
+      <VisualiserCreateNewReset />
     </Box>
   );
 };
